+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/GanttChart/Reflow/DynamicEventline.test.jsx
@@ -1,0 +1,83 @@
+import React from 'react';
+import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
+import Button from 'terra-button/lib/Button';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
+import graphData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/simplePanningData';
+import Carbon from '@cerner/carbon-graphs/src/js/carbon';
+
+const graphConfig = utils.deepClone(getGanttPanningConfig('#GanttPanningWithDynamicEventline'));
+graphConfig.eventline = [
+  {
+    color: Carbon.helpers.COLORS.GREY,
+    style: {
+      strokeDashArray: '4,4',
+    },
+    value: new Date(2016, 0, 1, 4).toISOString(),
+  },
+];
+
+const dataset = utils.deepClone(graphData);
+
+const initialState = {
+  initial: 0,
+  factor: 3,
+  dataset,
+  graphConfig,
+};
+
+const GanttPanningExample = () => {
+  const reducer = (panState, action) => {
+    const newGraphConfig = utils.deepClone(panState.graphConfig);
+    let hour;
+    const newDataset = [utils.deepClone(graphData[0])];
+
+    switch (action.type) {
+      case 'panLeft':
+        hour = panState.initial - panState.factor;
+        break;
+      case 'panRight':
+        hour = panState.initial + panState.factor;
+        break;
+      default:
+        return panState;
+    }
+
+    newGraphConfig.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    newGraphConfig.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+
+    const newEventline = [
+      {
+        color: Carbon.helpers.COLORS.BLACK,
+        style: {
+          strokeDashArray: '2,2',
+        },
+        value: new Date(2016, 0, 1, 6, 30).toISOString(),
+      },
+    ];
+
+    return {
+      initial: hour,
+      factor: panState.factor,
+      dataset: {
+        panData: newDataset,
+        eventline: newEventline,
+      },
+      graphConfig: utils.deepClone(newGraphConfig),
+    };
+  };
+
+  const [panState, dispatch] = React.useReducer(reducer, initialState);
+
+  return (
+    <React.Fragment>
+      <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
+      <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
+      <div id="tooltip" className="initial-tooltip" />
+      <GanttChart graphID="GanttPanningWithDynamicEventline" graphConfig={panState.graphConfig} dataset={panState.dataset} />
+    </React.Fragment>
+  );
+};
+
+export default GanttPanningExample;