+++ packages/terra-graphs/CHANGELOG.md
@@ -16,6 +16,7 @@
   * Added reflow to the Paired result component.
   * Added reflow to the Bar component.
   * Added reflow to the Bubble component.
+  * Added reflow to the Gantt component.
 
 * Changed
   * Made changes to WDIO test url's according to terra rules.